backend/requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
mysql-connector-python==8.1.1
python-dotenv==1.0.0